techpack/dataipa: affine ipa irqs to little CPUs
Some of the ipa irqs are run on CPU0 only and a few are run on CPU4.
Instead of utilizing the prime or gold clusters, affine these
completely to little cluster.
Little cluster can easily handle these interrupts and manage them
in a power efficient manner rather than handing these to the gold
clusters.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Change-Id: Icefae900d63832b6f5e50ddf034dab284b6779af
Signed-off-by: TogoFire <togofire@mailfence.com>